+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/NestedTryDepthCheckTest.java
@@ -34,7 +34,9 @@ public class NestedTryDepthCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "coding" + File.separator + filename);
+                + "coding" + File.separator
+                + "nestedtrydepth" + File.separator
+                + filename);
     }
 
     @Test
