+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/nestedtrydepth/InputNestedTryDepth.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.nestedtrydepth;
 
 public class InputNestedTryDepth
 {
